Resolved intermittent issue of replacing the results directory
Completed the fix based on progress made in the commit b4fbd10273c113634b29ec59cdcd8469e1619a9d

The previous commit placed the mkdir statement in a try/except clause, but referenced the wrong exception title. This commit rcorrected that exception title and empalced safeguards to ensure any errors unrelated to "File Exists" get reaised.

+++ src/sim_cup_main.py
@@ -152,10 +152,14 @@ while continueloop:
                 # This try statement slows it down to finish removing
                 # the directory before trying again to make it.
                 os.mkdir(results_dirname)
-            except FileExistsError:
-                # Try deleting again
-                shutil.rmtree(results_dirname,ignore_errors=True)
-                os.mkdir(results_dirname)
+            except OSError as exc:
+                if exc.errno != errno.EEXIST:
+                    raise
+                else:
+                    # Try deleting again
+                    shutil.rmtree(results_dirname,ignore_errors=True)
+                    os.mkdir(results_dirname)
+                #
             #
         elif not os.path.isdir(results_dirname):
             # The directory doesn't exist, make it now
